requirements.txt
streamlit
pandas
numpy
plotly
networkx
pyvis
requests
streamlit-lottie
gTTS
google-generativeai
python-dotenv